# backend/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Avanzada API",
    "description": "REST API for managing support requests (students, staff, admin).",
    "contact": {
      "name": "Avanzada"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/requests": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "List requests",
        "description": "Lists requests applying optional filters. Students only see their own requests.",
        "operationId": "listRequests",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Requests found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Create a new request",
        "description": "Creates a new support request. Students can only create requests for themselves; admins may specify another requester.",
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Request created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "403": {
            "description": "User not allowed to create request"
          }
        }
      }
    },
    "/api/requests/{id}/close": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Close a request",
        "description": "Closes a request with a closure observation. Only ADMIN can close.",
        "operationId": "close",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "409": {
            "description": "Invalid state transition"
          },
          "200": {
            "description": "Request closed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "User not allowed to close"
          }
        }
      }
    },
    "/api/requests/{id}/classify": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Classify a request",
        "description": "Sets type and priority of a request. Only STAFF and ADMIN can classify.",
        "operationId": "classify",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "409": {
            "description": "Invalid state transition"
          },
          "200": {
            "description": "Request classified",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "User not allowed to classify"
          }
        }
      }
    },
    "/api/requests/{id}/attend": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Mark request as attended",
        "description": "Marks a request as attended, with optional observations. Only STAFF and ADMIN can attend.",
        "operationId": "attend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendRequestDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "409": {
            "description": "Invalid state transition"
          },
          "200": {
            "description": "Request attended",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User not allowed to attend"
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/api/requests/{id}/assign": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Assign a request to a staff member",
        "description": "Assigns a request to a specific staff user. Only STAFF and ADMIN can assign.",
        "operationId": "assign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "409": {
            "description": "Invalid state transition"
          },
          "200": {
            "description": "Request assigned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "User not allowed to assign"
          }
        }
      }
    },
...
    "/api/requests/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get request by id",
        "description": "Returns a single request by id. Students can only access their own requests.",
        "operationId": "getRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Request found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "Student trying to access another user's request"
          }
        }
      }
    },
    "/api/requests/{id}/summary": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get request summary (AI or fallback)",
        "description": "Returns a textual summary of the request and its history using AI when available, or a fallback summary otherwise.",
        "operationId": "getSummary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Summary generated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "Student trying to access another user's request"
          }
        }
      }
    },
    "/api/requests/{id}/history": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get request history",
        "description": "Returns the history entries for a request. Students can only access the history of their own requests.",
        "operationId": "getHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "History entries found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntryDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "Student trying to access another user's request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateRequestDto": {
        "required": [
          "channelId",
          "description",
          "requestTypeId",
          "requestedById"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "requestTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "channelId": {
            "type": "integer",
            "format": "int64"
          },
          "requestedById": {
            "type": "integer",
            "format": "int64"
          },
          "registeredAt": {
            "type": "string"
          }
        }
      },
      "RequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "registeredAt": {
            "type": "string"
          },
          "requestTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "requestTypeCode": {
            "type": "string"
          },
          "requestTypeName": {
            "type": "string"
          },
          "channelId": {
            "type": "integer",
            "format": "int64"
          },
          "channelCode": {
            "type": "string"
          },
          "channelName": {
            "type": "string"
          },
          "stateId": {
            "type": "integer",
            "format": "int64"
          },
          "stateCode": {
            "type": "string"
          },
          "stateName": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "priorityJustification": {
            "type": "string"
          },
          "requestedById": {
            "type": "integer",
            "format": "int64"
          },
          "requestedByIdentifier": {
            "type": "string"
          },
          "requestedByName": {
            "type": "string"
          },
          "assignedToId": {
            "type": "integer",
            "format": "int64"
          },
          "assignedToIdentifier": {
            "type": "string"
          },
          "assignedToName": {
            "type": "string"
          },
          "closureObservation": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "CloseRequestDto": {
        "required": [
          "closureObservation"
        ],
        "type": "object",
        "properties": {
          "closureObservation": {
            "type": "string"
          }
        }
      },
      "ClassifyRequestDto": {
        "required": [
          "priority",
          "requestTypeId"
        ],
        "type": "object",
        "properties": {
          "requestTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "priority": {
            "type": "string"
          },
          "priorityJustification": {
            "type": "string"
          }
        }
      },
      "AttendRequestDto": {
        "type": "object",
        "properties": {
          "observations": {
            "type": "string"
          }
        }
      },
      "AssignRequestDto": {
        "required": [
          "assignedToId"
        ],
        "type": "object",
        "properties": {
          "assignedToId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SummaryResponseDto": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "fromAi": {
            "type": "boolean"
          }
        }
      },
      "HistoryEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "requestId": {
            "type": "integer",
            "format": "int64"
          },
          "occurredAt": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userIdentifier": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "observations": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Bearer token obtained via /api/auth/login",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}